refactor(vuls2): move the warning-entry helpers into fold.go
foldDetectionSeq's warnings harvest reached back into vuls2.go for
warningEntry / collectCriteriaWarnings / mergeWarningEntries; the
harvest is the fold's own responsibility, so the cluster (including
renderWarningEntries, called by detect()) lives with it.
